+++ modules/System/views/spaces/index.php
@@ -14,7 +14,6 @@
 
             <div class="kiss-margin-large-bottom kiss-flex kiss-flex-middle">
                 <span class="kiss-size-4 kiss-text-bold"><?=t('Spaces')?></span>
-                <span class="kiss-badge kiss-margin-small-left">BETA</span>
 
                 <div class="kiss-flex-1"></div>